sauvegarde ok - a continuer

+++ srcs/requirements/mariadb/scripts/entrypoint.sh
@@ -22,6 +22,7 @@
 
 #cat .setup 2> /dev/null
 #if [ $? -ne 0 ]; then
+	# https://dev.mysql.com/doc/refman/8.0/en/mysqld-safe.html
 	usr/bin/mysqld_safe --datadir=/var/lib/mysql &
 
 	# Apply config
@@ -29,15 +30,16 @@
 	sed -i "s|.*bind-address\s*=.*|bind-address=0.0.0.0\nport=3306|g" /etc/my.cnf.d/mariadb-server.cnf
 	
 	# Config MariaDB with default bases
-	mysqladmin --wait=30 ping; 
+	if ! mysqladmin --wait=30 ping; then
 	#then
-	#	printf "Unable to reach mariadb\n"
-	#	exit 1
-	#fi
+		printf "Unable to reach mariadb\n"
+		exit 1
+	fi
 
 	eval "echo \"$(cat /tmp/create_db.sql)\"" | mariadb
 	pkill mariadb
 	#touch .setup
 #fi
 
+#Sera utilise pour que le container wordpress puisse proceder a la configuration
 usr/bin/mysqld_safe --datadir=/var/lib/mysql
